VaultLib interface cleanups: remove .cipher_name, move .version
As noted in the comments in PR #12112, the cipher_name stored in the
VaultLib conflates the desired cipher for encryption with the cipher
read from an encrypted file header. This change makes the encryption
cipher a parameter to encrypt() and changes _split_header() to return
the extracted values so that decrypt() can use the appropriate cipher
name (which a new helper transforms into a VaultXXX class).

As a consequence, _split_header is renamed to _parse_vault_file, which
also becomes more strict about parsing vault headers. (This also means
that decrypt() doesn't have to check is_encrypted() separately, since
the validation is being done anyway.)

We also move b_version from VaultLib to the individual cipher classes so
that they can manage their own versioning; we also pass the version from
the file header to the cipher's decrypt() method.

This doesn't change any functionality, but it paves the way for easier
format changes without having to introduce ever more cipher classes.